Add files via upload
This code is designed to send a birthday message to a friend if it happens to be their birthday today. I employed Object-Oriented Programming (OOP) by representing my friends as individual objects. Within this setup, I created a method named "send" that handles the action of sending a "Happy Birthday" message to the friend but only does so if it's their birthday on the current day.